refactor: rename manager package to jobs
* and update related types and functions
