+++ internal/api/api.go
@@ -18,14 +18,13 @@ import (
 	"github.com/nais/v13s/internal/database"
 	"github.com/nais/v13s/internal/job"
 	"github.com/nais/v13s/internal/kubernetes"
-	"github.com/nais/v13s/internal/manager"
+	manager "github.com/nais/v13s/internal/management"
 	"github.com/nais/v13s/internal/metrics"
 	"github.com/nais/v13s/internal/model"
 	"github.com/nais/v13s/internal/sources"
 	"github.com/nais/v13s/internal/updater"
 	"github.com/nais/v13s/pkg/api/vulnerabilities"
 	"github.com/nais/v13s/pkg/api/vulnerabilities/management"
-	"github.com/prometheus/client_golang/prometheus"
 	"github.com/sirupsen/logrus"
 	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 	"golang.org/x/sync/errgroup"
@@ -54,13 +53,7 @@ func Run(ctx context.Context, cfg *config.Config, log logrus.FieldLogger) error 
 		Deleted: make(chan *model.Workload, 10000),
 	}
 
-	gFunc := prometheus.NewGaugeFunc(prometheus.GaugeOpts{
-		Name: "v13s_workload_update_queue_length",
-	}, func() float64 {
-		return float64(len(workloadEventQueue.Updated))
-	})
-
-	_, tp, promReg, err := metrics.NewMeterProvider(ctx, gFunc)
+	_, tp, promReg, err := metrics.NewMeterProvider(ctx)
 	if err != nil {
 		return fmt.Errorf("create metric meter: %w", err)
 	}
@@ -79,11 +72,11 @@ func Run(ctx context.Context, cfg *config.Config, log logrus.FieldLogger) error 
 		log.Fatalf("Failed to create verifier: %v", err)
 	}
 
-	jobCfg := &job.Config{
+	jobOpts := &job.Options{
 		DbUrl: cfg.DatabaseUrl,
 	}
 
-	mgr := manager.NewWorkloadManager(ctx, pool, jobCfg, verifier, source, workloadEventQueue, log.WithField("subsystem", "manager"))
+	mgr := manager.NewWorkloadManager(ctx, pool, jobOpts, verifier, source, workloadEventQueue, log.WithField("subsystem", "manager"))
 	mgr.Start(ctx)
 	defer mgr.Stop(ctx)
 
@@ -127,7 +120,7 @@ func Run(ctx context.Context, cfg *config.Config, log logrus.FieldLogger) error 
 			cfg.InternalListenAddr,
 			promReg,
 			log,
-			Handler{"/riverui", riverUI(ctx, jobCfg.DbUrl)},
+			Handler{"/riverui", riverUI(ctx, jobOpts.DbUrl)},
 		)
 	})
