+++ internal/api/grpcmgmt/server.go
@@ -11,7 +11,7 @@ import (
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/nais/v13s/internal/collections"
 	"github.com/nais/v13s/internal/database/sql"
-	"github.com/nais/v13s/internal/manager"
+	"github.com/nais/v13s/internal/jobs"
 	"github.com/nais/v13s/internal/model"
 	"github.com/nais/v13s/internal/updater"
 	"github.com/nais/v13s/pkg/api/vulnerabilities/management"
@@ -26,13 +26,13 @@ var _ management.ManagementServer = (*Server)(nil)
 type Server struct {
 	management.UnimplementedManagementServer
 	querier   sql.Querier
-	mgr       *manager.WorkloadManager
+	mgr       jobs.WorkloadManager
 	updater   *updater.Updater
 	parentCtx context.Context
 	log       *logrus.Entry
 }
 
-func NewServer(parentCtx context.Context, pool *pgxpool.Pool, mgr *manager.WorkloadManager, updater *updater.Updater, field *logrus.Entry) *Server {
+func NewServer(parentCtx context.Context, pool *pgxpool.Pool, mgr jobs.WorkloadManager, updater *updater.Updater, field *logrus.Entry) *Server {
 	return &Server{
 		parentCtx: parentCtx,
 		querier:   sql.New(pool),
