+++ internal/jobs/sbom/get.go
@@ -1,4 +1,4 @@
-package manager
+package sbom
 
 import (
 	"context"
@@ -6,10 +6,11 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/jackc/pgx/v5/pgtype"
 	"github.com/nais/v13s/internal/attestation"
 	"github.com/nais/v13s/internal/database/sql"
-	"github.com/nais/v13s/internal/job"
+	"github.com/nais/v13s/internal/jobs"
+	"github.com/nais/v13s/internal/jobs/output"
+	"github.com/nais/v13s/internal/jobs/types"
 	"github.com/nais/v13s/internal/model"
 	"github.com/riverqueue/river"
 	"github.com/sigstore/cosign/v2/pkg/cosign"
@@ -20,44 +21,20 @@ import (
 	"go.opentelemetry.io/otel/metric"
 )
 
-const (
-	KindGetAttestation = "get_attestation"
-)
-
-type GetAttestationJob struct {
-	ImageName    string
-	ImageTag     string
-	WorkloadId   pgtype.UUID
-	WorkloadType model.WorkloadType
-}
-
-func (GetAttestationJob) Kind() string { return KindGetAttestation }
-
-func (g GetAttestationJob) InsertOpts() river.InsertOpts {
-	return river.InsertOpts{
-		Queue: KindGetAttestation,
-		UniqueOpts: river.UniqueOpts{
-			ByArgs:   true,
-			ByPeriod: 1 * time.Minute,
-		},
-		MaxAttempts: 4,
-	}
-}
-
 type GetAttestationWorker struct {
-	db              sql.Querier
-	jobClient       job.Client
-	verifier        attestation.Verifier
-	workloadCounter metric.Int64UpDownCounter
-	log             logrus.FieldLogger
-	river.WorkerDefaults[GetAttestationJob]
+	Manager         jobs.WorkloadManager
+	Querier         sql.Querier
+	Verifier        attestation.Verifier
+	WorkloadCounter metric.Int64UpDownCounter
+	Log             logrus.FieldLogger
+	river.WorkerDefaults[types.GetAttestationJob]
 }
 
-func (g *GetAttestationWorker) NextRetry(job *river.Job[GetAttestationJob]) time.Time {
+func (g *GetAttestationWorker) NextRetry(job *river.Job[types.GetAttestationJob]) time.Time {
 	return time.Now().Add(1 * time.Minute)
 }
 
-func (g *GetAttestationWorker) Work(ctx context.Context, job *river.Job[GetAttestationJob]) error {
+func (g *GetAttestationWorker) Work(ctx context.Context, job *river.Job[types.GetAttestationJob]) error {
 	ctx, span := otel.Tracer("v13s/get-attestation").Start(ctx, "GetAttestationWorker.Work")
 	defer span.End()
 
@@ -70,14 +47,14 @@ func (g *GetAttestationWorker) Work(ctx context.Context, job *river.Job[GetAttes
 	imageName := job.Args.ImageName
 	imageTag := job.Args.ImageTag
 
-	g.log.WithFields(logrus.Fields{
+	g.Log.WithFields(logrus.Fields{
 		"image":      imageName,
 		"tag":        imageTag,
 		"workloadId": job.Args.WorkloadId.String(),
 	}).Debugf("getting attestation")
 
-	att, err := g.verifier.GetAttestation(ctx, fmt.Sprintf("%s:%s", imageName, imageTag))
-	g.workloadCounter.Add(
+	att, err := g.Verifier.GetAttestation(ctx, fmt.Sprintf("%s:%s", imageName, imageTag))
+	g.WorkloadCounter.Add(
 		ctx,
 		1,
 		metric.WithAttributes(
@@ -92,10 +69,9 @@ func (g *GetAttestationWorker) Work(ctx context.Context, job *river.Job[GetAttes
 			if err.Error() != "no matching attestations: " {
 				span.RecordError(err)
 				span.SetStatus(codes.Error, err.Error())
-				g.log.WithError(err).Error("failed to get attestation")
+				g.Log.WithError(err).Error("failed to get attestation")
 			}
-			// TODO: handle errors
-			err = g.db.UpdateWorkloadState(ctx, sql.UpdateWorkloadStateParams{
+			err = g.Querier.UpdateWorkloadState(ctx, sql.UpdateWorkloadStateParams{
 				State: sql.WorkloadStateNoAttestation,
 				ID:    job.Args.WorkloadId,
 			})
@@ -103,7 +79,7 @@ func (g *GetAttestationWorker) Work(ctx context.Context, job *river.Job[GetAttes
 				return fmt.Errorf("failed to set workload state: %w", err)
 			}
 
-			err = g.db.UpdateImageState(ctx, sql.UpdateImageStateParams{
+			err = g.Querier.UpdateImageState(ctx, sql.UpdateImageStateParams{
 				State: sql.ImageStateFailed,
 				Name:  imageName,
 				Tag:   imageTag,
@@ -111,7 +87,7 @@ func (g *GetAttestationWorker) Work(ctx context.Context, job *river.Job[GetAttes
 			if err != nil {
 				return fmt.Errorf("failed to set image state: %w", err)
 			}
-			recordOutput(ctx, JobStatusNoAttestation)
+			output.Record(ctx, output.JobStatusNoAttestation)
 			if job.Args.WorkloadType == model.WorkloadTypeApp {
 				return noMatchAttestationError
 			}
@@ -119,16 +95,16 @@ func (g *GetAttestationWorker) Work(ctx context.Context, job *river.Job[GetAttes
 		} else if errors.As(err, &unrecoverableError) {
 			span.RecordError(err)
 			span.SetStatus(codes.Error, err.Error())
-			g.log.WithError(err).Error("unrecoverable error while getting attestation")
-			recordOutput(ctx, JobStatusUnrecoverable)
-			err = g.db.UpdateWorkloadState(ctx, sql.UpdateWorkloadStateParams{
+			g.Log.WithError(err).Error("unrecoverable error while getting attestation")
+			output.Record(ctx, output.JobStatusUnrecoverable)
+			err = g.Querier.UpdateWorkloadState(ctx, sql.UpdateWorkloadStateParams{
 				State: sql.WorkloadStateUnrecoverable,
 				ID:    job.Args.WorkloadId,
 			})
 			if err != nil {
 				return fmt.Errorf("failed to set workload state: %w", err)
 			}
-			err = g.db.UpdateImageState(ctx, sql.UpdateImageStateParams{
+			err = g.Querier.UpdateImageState(ctx, sql.UpdateImageStateParams{
 				State: sql.ImageStateFailed,
 				Name:  imageName,
 				Tag:   imageTag,
@@ -138,7 +114,7 @@ func (g *GetAttestationWorker) Work(ctx context.Context, job *river.Job[GetAttes
 			}
 			return river.JobCancel(unrecoverableError)
 		} else {
-			return handleJobErr(err)
+			return output.HandleJobErr(err)
 		}
 	}
 	if att != nil {
@@ -146,7 +122,7 @@ func (g *GetAttestationWorker) Work(ctx context.Context, job *river.Job[GetAttes
 		if err != nil {
 			return fmt.Errorf("failed to compress attestation: %w", err)
 		}
-		err = g.jobClient.AddJob(ctx, &UploadAttestationJob{
+		err = g.Manager.AddJob(ctx, &types.UploadAttestationJob{
 			ImageName:   imageName,
 			ImageTag:    imageTag,
 			WorkloadId:  job.Args.WorkloadId,
@@ -157,7 +133,7 @@ func (g *GetAttestationWorker) Work(ctx context.Context, job *river.Job[GetAttes
 			span.SetStatus(codes.Error, err.Error())
 			return err
 		}
-		recordOutput(ctx, JobStatusAttestationDownloaded)
+		output.Record(ctx, output.JobStatusAttestationDownloaded)
 	}
 	return nil
 }
