+++ internal/jobs/types/upload_attestation.go
@@ -1,0 +1,33 @@
+package types
+
+import (
+	"time"
+
+	"github.com/jackc/pgx/v5/pgtype"
+	"github.com/riverqueue/river"
+)
+
+const (
+	KindUploadAttestation            = "upload_attestation"
+	UploadAttestationByPeriodMinutes = 2 * time.Minute
+)
+
+type UploadAttestationJob struct {
+	ImageName   string `river:"unique"`
+	ImageTag    string `river:"unique"`
+	WorkloadId  pgtype.UUID
+	Attestation []byte
+}
+
+func (UploadAttestationJob) Kind() string { return KindUploadAttestation }
+
+func (u UploadAttestationJob) InsertOpts() river.InsertOpts {
+	return river.InsertOpts{
+		Queue: KindUploadAttestation,
+		UniqueOpts: river.UniqueOpts{
+			ByArgs:   true,
+			ByPeriod: UploadAttestationByPeriodMinutes,
+		},
+		MaxAttempts: 8,
+	}
+}
